workspace/swag-latent-diffusion/requirements.txt
absl-py==2.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
Deprecated==1.2.14
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
glob2==0.7
grpcio==1.62.0
humanize==4.9.0
idna==3.6
imageio==2.34.0
Jinja2==3.1.3
kiwisolver==1.4.5
lazy_loader==0.3
lpips==0.1.4
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
monai-weekly==1.2.dev2304
mpmath==1.3.0
networkx==3.2.1
nibabel==5.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
onnx==1.15.0
onnx2torch==1.5.13
packaging==23.2
pandas==2.2.1
pillow==10.2.0
protobuf==4.25.3
pydicom==2.4.4
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
rich==13.7.0
scikit-image==0.22.0
scipy==1.12.0
shellingham==1.5.4
SimpleITK==2.3.1
six==1.16.0
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tifffile==2024.2.12
torch==2.2.1
torchio==0.19.6
torchvision==0.17.1
torchxrayvision==1.2.2
tqdm==4.66.2
triton==2.2.0
typer==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
wrapt==1.16.0